+++ sys/arch/mvme68k/mvme68k/disksubr.c
@@ -1,4 +1,4 @@
-/*	$NetBSD: disksubr.c,v 1.22 2002/02/19 17:09:46 wiz Exp $	*/
+/*	$NetBSD: disksubr.c,v 1.23 2002/03/05 09:40:40 simonb Exp $	*/
 
 /*
  * Copyright (c) 1995 Dale Rahn.
@@ -147,7 +147,7 @@ setdisklabel(olp, nlp, openmask, clp)
 	    dkcksum(nlp) != 0)
 		return (EINVAL);
 
-	while ((i = ffs((long)openmask)) != 0) {
+	while ((i = ffs(openmask)) != 0) {
 		i--;
 		openmask &= ~(1 << i);
 		if (nlp->d_npartitions <= i)
